backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Maketribe Designs Creation"
authors = ["Fred Bone"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
openai = "^1.29.0"
python-dotenv = "^0.19.2"


torch = "^2.3.1"
torchvision = "^0.18.1"
transformers = "^4.41.2"
pillow = "^10.3.0"
numpy = "^2.0.0"
aiofiles = "^24.1.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"